create Particle

+++ XGF/Src/Core/FPS.cpp
@@ -36,7 +36,7 @@ void FPS::RenderFPS(TextRenderer & TextRenderer, int x, int y)
 {
 	wchar_t fps[30];
 	swprintf(fps, 29, L"FPS:%.2f", mFPS);
-	TextRenderer.DrawString(fps, x,y);
+	TextRenderer.DrawString(fps, static_cast<float>(x),static_cast<float>(y));
 }
 
 void FPS::RenderFPS(TextRenderer & TextRenderer, Color color, int x, int y)
